Extract EntityBoxCamera class.
Now there's a nice way to encapsulate the camera logic.
Also there's a way to draw on the screen using absolute coodinates,
which reduces a bunch of math.
Also fixed a nasty performance-killing bug: the background was being
drawn with floating point coordinates, which made desktop browsers run
really slow!